Refactored SentencePiece Vocab and Models, removed duplicate "populate_masks" calls
