+++ main/src/tokenizer/marian_tokenizer.rs
@@ -173,7 +173,7 @@ impl Tokenizer<MarianVocab> for MarianTokenizer {
             }
             is_prev_unknown = node.index == 0;
         }
-        SentencePieceModel::populate_masks(output.as_mut_slice(), '\u{2581}');
+        self.model.populate_masks(output.as_mut_slice(), '\u{2581}');
         output
     }
