+++ main/src/vocab/mod.rs
@@ -45,6 +45,8 @@ mod pegasus_vocab;
 mod prophetnet_vocab;
 mod reformer_vocab;
 mod roberta_vocab;
+mod sentence_piece_bpe_model;
+mod sentence_piece_unigram_model;
 mod sentence_piece_vocab;
 pub(crate) mod sentencepiece_proto;
 mod t5_vocab;
@@ -64,7 +66,9 @@ pub use pegasus_vocab::PegasusVocab;
 pub use prophetnet_vocab::ProphetNetVocab;
 pub use reformer_vocab::ReformerVocab;
 pub use roberta_vocab::RobertaVocab;
-pub use sentence_piece_vocab::{SentencePieceModel, SentencePieceVocab};
+pub use sentence_piece_bpe_model::SentencePieceBpeModel;
+pub use sentence_piece_unigram_model::SentencePieceModel;
+pub use sentence_piece_vocab::SentencePieceVocab;
 pub use t5_vocab::T5Vocab;
 pub use xlm_roberta_vocab::XLMRobertaVocab;
 pub use xlnet_vocab::XLNetVocab;
